askme-backend/requirements.txt
flask
python-dotenv
langchain
langchain-google-genai
google-genai
mem0ai
flask-cors

livekit-agents[turn-detector]~=1.2
livekit-agents
livekit-plugins-google
livekit-plugins-silero
livekit-api
livekit-plugins-noise-cancellation~=0.2
livekit-agents[sarvam]~=1.2
livekit-plugins-sarvam
livekit-agents[groq]~=1.2

#   stt=sarvam.STT(
#             language="hi-IN",
#             model="saarika:v2.5",
#             api_key=os.getenv("SARVAM_API_KEY"),
#         ),
#         tts= google.beta.GeminiTTS(
#                 model="gemini-2.5-flash-preview-tts",
#                 voice_name="Zephyr",
#                 instructions="Speak in a friendly and engaging tone.",
#             ),
#         llm=google.LLM(
#             model="gemini-1.5-flash",
#             temperature=0.7,
#             api_key=os.getenv("GOOGLE_API_KEY"),
#         ),